requirements.txt
# 核心爬虫依赖
requests>=2.28.0
beautifulsoup4>=4.11.0

# 数据处理
pandas>=1.5.0

# 数据库连接
psycopg2-binary>=2.9.0

# 定时任务
schedule>=1.2.0

# 可选依赖 (用于更好的性能和功能)
lxml>=4.9.0          # 更快的XML/HTML解析
python-dotenv>=0.19.0 # 环境变量管理